% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K2C.R
\name{kelvin_to_celsius}
\alias{kelvin_to_celsius}
\title{KC2}
\usage{
kelvin_to_celsius(temp_K)
}
\arguments{
\item{temp_K}{temperature in Kelvin}
}
\value{
a numerical value of the temperature in Celcius
}
\description{
KC2
}
\examples{
kelvin_to_celsius(400) # 126.85
}
